moved fastify tests to versioned folder so they properly run against the fastify versions. fixed the tests to work on v3
