+++ test/versioned/fastify/package.json
@@ -8,10 +8,14 @@
         "node": ">=12"
       },
       "dependencies": {
-        "fastify": ">=2.0.0"
+        "fastify": ">=2.0.0",
+        "middie": "5.3.0"
       },
       "files": [
-        "integration.tap.js"
+        "errors.tap.js",
+        "naming.tap.js",
+        "fastify.tap.js",
+        "new-state-tracking.tap.js"
       ]
     }
   ]
